# specs/001-whatsapp-drops/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: DropInDrop WhatsApp E-commerce API
  description: API for managing WhatsApp-based e-commerce operations including products, drops, orders, and delivery verification
  version: 1.0.0
  contact:
    name: DropInDrop API Support
    email: support@dropindrop.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.dropindrop.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\+[1-9]\d{1,14}$'
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, DELIVERY_MANAGER, CLIENT]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 5
          maxLength: 200
        description:
          type: string
        price:
          type: number
          minimum: 0
          maximum: 999999.99
        sku:
          type: string
          maxLength: 50
        categoryId:
          type: string
          format: uuid
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
        isActive:
          type: boolean
        stockQuantity:
          type: integer
          minimum: 0
        viewCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
          pattern: '^ORD-\d{8}-\d{4}$'
        customerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, PAID, DELIVERED, CANCELLED]
        totalAmount:
          type: number
          minimum: 0
        paymentIntentId:
          type: string
          maxLength: 100
        paidAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        deliveredBy:
          type: string
          format: uuid
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        qrCodeData:
          type: string
        uniqueCode:
          type: string
          pattern: '^[A-Z]{4}-\d{4}$'
        isUsed:
          type: boolean
        usedAt:
          type: string
          format: date-time
        usedBy:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        code:
          type: string
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
paths:
...
  /orders:
    get:
      summary: List orders
      description: Get orders with filtering and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PAID, DELIVERED, CANCELLED]
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          orders:
                            type: array
                            items:
                              $ref: "#/components/schemas/Order"
                          pagination:
                            type: object
                            properties:
                              page:
                                type: integer
                              limit:
                                type: integer
                              total:
                                type: integer
                              totalPages:
                                type: integer
  /orders/{id}:
    get:
      summary: Get order details
      description: Get detailed information about a specific order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/Order"
                          - type: object
                            properties:
                              customer:
                                $ref: "#/components/schemas/User"
                              items:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    product:
                                      $ref: "#/components/schemas/Product"
                                    quantity:
                                      type: integer
                                    unitPrice:
                                      type: number
                                    totalPrice:
                                      type: number
                              ticket:
                                $ref: "#/components/schemas/Ticket"

  # Ticket verification endpoints
  /tickets/verify:
    post:
      summary: Verify ticket for delivery
      description: Verify ticket using QR code data or unique code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qrCodeData:
                  type: string
                uniqueCode:
                  type: string
                  pattern: '^[A-Z]{4}-\d{4}$'
      responses:
        "200":
          description: Ticket verified successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ticket:
                            $ref: "#/components/schemas/Ticket"
                          order:
                            $ref: "#/components/schemas/Order"
                          customer:
                            $ref: "#/components/schemas/User"
        "400":
          description: Invalid ticket or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Client order endpoints
...